delete route now has auth and checks if user owns the Link

+++ backend/src/controllers/urlController.ts
@@ -2,6 +2,7 @@ import { Request, Response } from 'express';
 import { alphaNumSample } from '../utils/codeGen';
 import { Collection } from 'mongodb';
 import URLModel from '../models/urlModel';
+import User from '../models/userModel';
 
 const getUrlByCode = async (req: Request, res: Response) => {
     const { code } = req.params;
@@ -53,13 +54,19 @@ const shortenURL = async (req: Request, res: Response) => {
 
 const deleteURL = async (req: Request, res: Response) => {
     const { code } = req.params;
+    const user = res.locals.user as User;
+
+    if (!user.email) return res.status(401).json({ message: `Unauthorized` });
 
     try {
         const urlsCollection = req.app.get('urlsCollection') as Collection<URLModel>;
 
+        const url = await urlsCollection.findOne({ code: code });
+        if (!url) return res.status(404).json({ message: `URL not found` });
+        if (url.creator_email !== user.email) return res.status(401).json({ message: `Unauthorized` });
+
         const result = await urlsCollection.deleteOne({ code: code });
 
-        if (result.deletedCount === 0) return res.status(404).json({ message: `URL not found` });
         if (result.acknowledged === false) return res.status(500).json({ message: `Error deleting URL` });
 
         return res.status(200).json({ message: `URL deleted` });
